[TNT-269] refactor: mealrecord 모듈 내 식단 기록 상세 코드 삭제
