+++ feature/trainee/mealrecord/src/main/java/co/kr/tnt/trainee/mealrecord/navigation/TraineeMealRecordNavigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavOptionsBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import co.kr.tnt.navigation.Route
-import co.kr.tnt.trainee.mealrecord.record.TraineeMealRecordRoute
+import co.kr.tnt.trainee.mealrecord.TraineeMealRecordRoute
 
 fun NavController.navigateToTraineeMealRecord(
     selectedDate: String,
